modif supprProfil.php

+++ admin/supprProfil.php
@@ -1,29 +1,38 @@
 <?php
 require 'lib.inc.php';
 
-$email = $_POST['email'];
+session_start();
 
-try {
-    $mabd = connexionBD();
-    $req = $mabd->prepare('SELECT COUNT(*) as count FROM utilisateurs WHERE user_mail = :email');
-    $req->execute(array(':email' => $email));
-    $result = $req->fetch(PDO::FETCH_ASSOC);
+if (isset($_SESSION['user_mail'])) {
+    $email = $_SESSION['user_mail'];
 
-    if ($result['count'] > 0) {
-        //suppr l'utilisateur et ses données
-        $req = $mabd->prepare('DELETE FROM utilisateurs WHERE user_mail = :email');
+    try {
+        $mabd = connexionBD();
+        $req = $mabd->prepare('SELECT COUNT(*) as count FROM utilisateurs WHERE user_mail = :email');
         $req->execute(array(':email' => $email));
+        $result = $req->fetch(PDO::FETCH_ASSOC);
 
-        header('location: ../index.php?deleted=1');
-        exit();
-    } else {
-        echo "Utilisateur non trouvé !";
-        header('location: ../modifProfil.php?erreur=1');
-        exit();
+        if ($result['count'] > 0) {
+            // Supprimer l'utilisateur et ses données
+            $req = $mabd->prepare('DELETE FROM utilisateurs WHERE user_mail = :email');
+            $req->execute(array(':email' => $email));
+
+            session_destroy(); // Détruire la session après la suppression du compte
+
+            header('location: ../index.php?deleted=1');
+            exit();
+        } else {
+            echo "Utilisateur non trouvé !";
+            header('location: ../modifProfil.php?erreur=1');
+            exit();
+        }
+    } catch (PDOException $e) {
+        die("Erreur de connexion à la base de données : " . $e->getMessage());
     }
-} catch (PDOException $e) {
-    die("Erreur de connexion à la base de données : " . $e->getMessage());
-}
 
-deconnexionBD($mabd);
+    deconnexionBD($mabd);
+} else {
+    header('location: ../index.php');
+    exit();
+}
 ?>
